% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatetracker_func.R
\name{get_hatetracker_activity}
\alias{get_hatetracker_activity}
\title{Fetch trends with 30-minute resolution on a specific day.}
\usage{
get_hatetracker_activity(start_date = Sys.Date())
}
\arguments{
\item{start_date}{Beginning of date range.}
}
\description{
Fetch trends with 30-minute resolution on a specific day.
}
