% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatetracker_func.R
\name{get_hatetracker_activity2}
\alias{get_hatetracker_activity2}
\title{Fetch trends over a multi-day period with daily resolution.}
\usage{
get_hatetracker_activity2(start_date = Sys.Date(), end_date = NULL)
}
\arguments{
\item{start_date}{Beginning of date range.}

\item{end_date}{End of date range.}
}
\description{
Fetch trends over a multi-day period with daily resolution.
}
