% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatetracker.R
\docType{package}
\name{hatetracker}
\alias{hatetracker}
\alias{hatetracker-package}
\title{hatetracker: A package for working with the SPLC hatetracker.io API.}
\description{
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{hatetracker functions}{

The foo functions ...
}

