% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hatetracker_func.R
\name{get_hatetracker_activity}
\alias{get_hatetracker_activity}
\title{Fetch current trends}
\usage{
get_hatetracker_activity(start_date = Sys.Date(), end_date = NULL)
}
\arguments{
\item{start_date}{Beginning of date range.}

\item{end_date}{End of date range.}
}
\description{
Fetch current trends
}
